expose seperate read module and improve panic safety
This commit moves RunState and RleDecoder to a seperate module
to prevent name conflicts in the future.
Furthermore, it updates states eagerly to not forget them
if the unerlying reader panics.
